[swing] Extract separate SwingRedrawer interface
